split ResultsView
